+++ kmath-core/src/commonTest/kotlin/space/kscience/kmath/testutils/AlgebraicVerifier.kt
@@ -1,0 +1,9 @@
+package space.kscience.kmath.testutils
+
+import space.kscience.kmath.operations.Algebra
+
+internal interface AlgebraicVerifier<T, out A> where A : Algebra<T> {
+    val algebra: A
+
+    fun verify()
+}