+++ kmath-geometry/src/commonMain/kotlin/space/kscience/kmath/geometry/GeometrySpace.kt
@@ -1,0 +1,17 @@
+package space.kscience.kmath.geometry
+
+import space.kscience.kmath.operations.Space
+
+public interface Vector
+
+public interface GeometrySpace<V : Vector> : Space<V> {
+    /**
+     * L2 distance
+     */
+    public fun V.distanceTo(other: V): Double
+
+    /**
+     * Scalar product
+     */
+    public infix fun V.dot(other: V): Double
+}