+++ kmath-histograms/src/jvmMain/kotlin/space/kscience/kmath/histogram/Counters.kt
@@ -1,0 +1,7 @@
+package space.kscience.kmath.histogram
+
+import java.util.concurrent.atomic.DoubleAdder
+import java.util.concurrent.atomic.LongAdder
+
+public actual typealias LongCounter = LongAdder
+public actual typealias DoubleCounter = DoubleAdder